Created new types for devices. #179
Devices were expressed with a single record "Device",
which had the `Kind` field in it to show which kind of device it is,
and the rest of the field were the same among all the kind.
Now devices are expressed with specified kinds of record,
`Client`, `Router` and `Hub`.
This change makes it possible to see which device is of which kind,
and to let devices have fields that represents the device better.