Removed some classes with a single property for each. #218
Some classes had a single property for each,
such as `font-weight: 400;`
though its name made me feel the class was for a special purpose,
such as `column-addition-area`.
Now these classes are replaced with ones with a sigle property for each
and a simple name which tells how it works, such as `regular`.