Fixed an issue with the Complement page. Close #329.
Hint messages in the page were wrong. The issue was solved by adding a new parameter to `EndlessBinary.exec` to convert a randomly generated number to the answer of a question.
In most pages, the random number (let's say it's 2) becomes the answer ("10") directly, so the number (2) will be displayed on the screen and compared with the user input.
But in the Complement page, the answer is the complement of the random number, so if our random number is 2 ("0010"), the answer is "1110", so we need a conversion function.

Addition, Subtraction, and NotFound pages have their unique logic, so they don't share code with `EndlessBinary.exec`.